backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
gitpython
langchain
langchain-community
langchain-google-genai
langchain-chroma
langchain-text-splitters
langchain-core
chromadb
PyGithub
sqlalchemy
pytest
aiosqlite
sentence-transformers